requirements.txt
langchain
torch
accelerate
sentence_transformers
streamlit_chat
streamlit
faiss-cpu
tiktoken
huggingface-hub
pypdf
# llama-cpp-python
# langchain-community # pip install -U langchain-community